refactor: 운동 추천 컨트롤러 분리
